backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pymongo>=4.6.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
